requirements.txt
pycryptodome>=3.10
pyoqs>=0.7.0 ; platform_system != "Windows"
pytest>=7.0
pytest-cov